Introducing base implemention for Events
