+++ src/Event/ErrorEvent.php
@@ -2,19 +2,6 @@
 
 namespace Phlow\Event;
 
-use Phlow\Workflow\WorkflowStep;
+class ErrorEvent extends BaseEvent {
 
-class ErrorEvent implements Event
-{
-    private $nextStep;
-
-    public function __construct(WorkflowStep $nextStep)
-    {
-        $this->nextStep = $nextStep;
-    }
-
-    public function next()
-    {
-        return $this->nextStep;
-    }
 }
