+++ src/Event/StartEvent.php
@@ -2,19 +2,7 @@
 
 namespace Phlow\Event;
 
-use Phlow\Workflow\WorkflowStep;
-
-class StartEvent implements Event
+class StartEvent extends BaseEvent
 {
-    private $nextStep;
 
-    public function __construct(WorkflowStep $nextStep)
-    {
-        $this->nextStep = $nextStep;
-    }
-
-    public function next()
-    {
-        return $this->nextStep;
-    }
 }
